+++ crates/libwebrtc-sys/include/adm.h
@@ -70,7 +70,7 @@ class ExtendedADM : public webrtc::AudioDeviceModule {
   // Stops the `bridge::LocalAudioSource` for the provided device ID.
   virtual void DisposeAudioSource(std::string device_id) = 0;
 
-  // Returns inner `PlayoutDelegatingAPM`.
+  // Returns the inner `PlayoutDelegatingAPM`.
   virtual rtc::scoped_refptr<PlayoutDelegatingAPM> AudioProcessing() = 0;
 };
 
@@ -96,7 +96,7 @@ class OpenALAudioDeviceModule : public ExtendedADM {
   // Stops the `bridge::LocalAudioSource` for the provided device ID.
   void DisposeAudioSource(std::string device_id) override;
 
-  // Returns `PlayoutDelegatingAPM` used by this `OpenALAudioDeviceModule`.
+  // Returns the `PlayoutDelegatingAPM` used by this `OpenALAudioDeviceModule`.
   rtc::scoped_refptr<PlayoutDelegatingAPM> AudioProcessing() override;
 
   // Playout control.